Add style/indent :defn to hooks that have deps.
As part of this change I moved where the reader conditional was. Before this, things like cider could not find the definition of these macros in cljs, so they couldn't pull the metadata. That means indention didn't work, neither did function signature help or docstrings. Now all of these things work, but we also ensure that the body is still only executed on the clj side.